Switched to MutationEvent.TICK
### MutationEvent ###
Now controlls the main events, giving additional functionality for pausing segments of the game, making things play every second frame, or just know the tickCount etc.
In future, there should be a main game, a constant tick and an overlay tick that run simultaneously - so certain aspects can be paused while others are not.